+++ duplicates/cli.py
@@ -2,19 +2,20 @@ import logging
 import os
 from pathlib import Path
 import sys
-from click import argument, command, confirm, option
+from click import Abort, argument, command, confirm, option, version_option
 
-from . import DuplicateFiles, LOGROOT
+from . import DuplicateFiles, InvalidFileSystemError, LOGROOT
 
 _logger = logging.getLogger(LOGROOT)
 
 @command()
-@argument('rootdir')
+@argument('rootdirs', nargs=-1)
 @option('--link', is_flag=True)
 @option('-y', 'approved', is_flag=True)
 @option('--list', '_list', is_flag=True)
 @option('--short', is_flag=True)
-def dupes(rootdir, link, approved, _list, short):
+@version_option(package_name='link_duplicates')
+def dupes(rootdirs, link, approved, _list, short):
     _logger.setLevel(logging.INFO)
     consoleoutput = logging.StreamHandler()
     consoleoutput.setLevel(logging.INFO)
@@ -23,8 +24,13 @@ def dupes(rootdir, link, approved, _list, short):
     consoleoutput.setFormatter(outputformat)
     _logger.addHandler(consoleoutput)
 
-    rootdir = Path(rootdir)
-    duplicatefiles = DuplicateFiles.frompath(rootdir)
+    rootdirs = [Path(rootdir) for rootdir in rootdirs]
+    
+    try:
+        duplicatefiles = DuplicateFiles.frompaths(*rootdirs)
+    except InvalidFileSystemError:
+        print(f'{', '.join(os.fspath(p) for p in rootdirs)} are not on the same filesystem')
+        raise Abort
     
     if short:
         print(duplicatefiles.printout(ignoresamenames=True))
@@ -35,6 +41,6 @@ def dupes(rootdir, link, approved, _list, short):
         if not approved:
             confirm('Link files?', abort=True, err=True) 
             print('', file=sys.stderr) #prompting to stderr doesn't echo input (including \n)
-        _logger.info(f'Linking files in {os.fspath(rootdir)} ...')
+        _logger.info(f'Linking files in {', '.join(os.fspath(p) for p in rootdirs)} ...')
         duplicatefiles.link()
         _logger.info(f'Done')